Correction bug

+++ src/controllers/articleController.js
@@ -31,30 +31,78 @@ exports.deleteArticle = (req, res, next) => {
   if (req.userRole !== "admin") {
     const message = "Vous n'avez pas les droits pour créer un article";
     return res.status(401).json({ message });
+  } else {
+    Article.findOne({ where: { id: req.params.id } })
+      .then((article) => {
+        fs.unlink(__basedir + "/" + article.img_url, (err) => {
+          if (err) {
+            const message =
+              "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
+            return res.status(500).json({ message, data: err });
+          }
+        });
+        Article.destroy({ where: { id: req.params.id } })
+          .then((article) => {
+            const message = "L'article a été supprimé avec succès";
+            return res.json({ message });
+          })
+          .catch((error) => {
+            const message =
+              "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
+            return res.json({ message, data: error });
+          });
+      })
+      .catch((error) => {
+        const message =
+          "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
+        return res.status(500).json({ message, data: error });
+      });
   }
+};
+
+exports.updateArticle = (req, res, next) => {
+  if (req.userRole !== "admin") {
+    const message = "Vous n'avez pas les droits pour créer un article";
+    return res.status(401).json({ message });
+  }
+  let imagePath = req.body.img_url;
+
+  if (req.file) {
+    imagePath = req.file.path;
+  }
+
   Article.findOne({ where: { id: req.params.id } })
     .then((article) => {
-      fs.unlink(__basedir + "/" + article.img_url, (err) => {
-        if (err) {
-          const message =
-            "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
-          return res.status(500).json({ message, data: err });
-        }
-      });
-      Article.destroy({ where: { id: req.params.id } })
+      if (imagePath !== article.img_url) {
+        fs.unlink(__basedir + "/" + article.img_url, (err) => {
+          if (err) {
+            const message =
+              "L'article n'a pas pu être modifié. Réessayez dans quelques instants";
+            return res.status(500).json({ message, data: err });
+          }
+        });
+      }
+      Article.update(
+        {
+          title: req.body.title,
+          content: req.body.content,
+          img_url: imagePath,
+        },
+        { where: { id: req.params.id } }
+      )
         .then((article) => {
-          const message = "L'article a été supprimé avec succès";
-          return res.json({ message });
+          const message = "L'article a été modifié avec succès";
+          return res.json({ message, data: article });
         })
         .catch((error) => {
           const message =
-            "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
+            "L'article n'a pas pu être modifié. Réessayez dans quelques instants";
           return res.json({ message, data: error });
         });
     })
     .catch((error) => {
       const message =
-        "L'article n'a pas pu être supprimé. Réessayez dans quelques instants";
+        "L'article n'a pas pu être modifié. Réessayez dans quelques instants";
       return res.status(500).json({ message, data: error });
     });
 };
